requirement.txt
streamlit
numpy
opencv-python
pillow
joblib
nilearn
scikit-learn
fpdf
matplotlib